+++ src/include/smash/propagation.h
@@ -100,12 +100,14 @@ void expand_space_time(Particles *particles,
  * \param[in] FI3_lat Lattice for the electric and magnetic
  *            components of the symmetry force
  * \param[in] EM_lat Lattice for the electric and magnetic field
+ * \param[in] jB_lat Lattice of the net baryon density
  */
 void update_momenta(
     std::vector<Particles> &particles, double dt, const Potentials &pot,
     RectangularLattice<std::pair<ThreeVector, ThreeVector>> *FB_lat,
     RectangularLattice<std::pair<ThreeVector, ThreeVector>> *FI3_lat,
-    RectangularLattice<std::pair<ThreeVector, ThreeVector>> *EM_lat);
+    RectangularLattice<std::pair<ThreeVector, ThreeVector>> *EM_lat,
+    DensityLattice *jB_lat);
 
 }  // namespace smash
 #endif  // SRC_INCLUDE_SMASH_PROPAGATION_H_