+++ src/propagation.cc
@@ -112,7 +112,8 @@ void update_momenta(
     std::vector<Particles> &ensembles, double dt, const Potentials &pot,
     RectangularLattice<std::pair<ThreeVector, ThreeVector>> *FB_lat,
     RectangularLattice<std::pair<ThreeVector, ThreeVector>> *FI3_lat,
-    RectangularLattice<std::pair<ThreeVector, ThreeVector>> *EM_lat) {
+    RectangularLattice<std::pair<ThreeVector, ThreeVector>> *EM_lat,
+    DensityLattice *jB_lat) {
   // Copy particles from ALL ensembles to a single list before propagation
   // and calculate potentials from this list
   ParticleList plist;
@@ -155,25 +156,39 @@ void update_momenta(
         FB = std::make_pair(std::get<0>(tmp), std::get<1>(tmp));
         FI3 = std::make_pair(std::get<2>(tmp), std::get<3>(tmp));
       }
-      ThreeVector Force =
-          scale.first *
-              (FB.first + data.momentum().velocity().cross_product(FB.second)) +
-          scale.second * data.type().isospin3_rel() *
-              (FI3.first +
-               data.momentum().velocity().cross_product(FI3.second));
+      /* Floating point traps should be raised if the force is not overwritten
+       * with a meaningful value */
+      const auto sNaN = std::numeric_limits<double>::signaling_NaN();
+      ThreeVector force(sNaN, sNaN, sNaN);
+      if (pot.use_momentum_dependence()) {
+        ThreeVector energy_grad = pot.single_particle_energy_gradient(
+            jB_lat, data.position().threevec(), data.momentum().threevec(),
+            data.effective_mass(), plist);
+        force = -energy_grad * scale.first;
+        force +=
+            scale.second * data.type().isospin3_rel() *
+            (FI3.first + data.momentum().velocity().cross_product(FI3.second));
+      } else {
+        force = scale.first *
+                    (FB.first +
+                     data.momentum().velocity().cross_product(FB.second)) +
+                scale.second * data.type().isospin3_rel() *
+                    (FI3.first +
+                     data.momentum().velocity().cross_product(FI3.second));
+      }
       // Potentially add Lorentz force
       if (pot.use_coulomb() && EM_lat->value_at(r, EM_fields)) {
         // factor hbar*c to convert fields from 1/fm^2 to GeV/fm
-        Force += hbarc * data.type().charge() * elementary_charge *
+        force += hbarc * data.type().charge() * elementary_charge *
                  (EM_fields.first +
                   data.momentum().velocity().cross_product(EM_fields.second));
       }
-      logg[LPropagation].debug("Update momenta: F [GeV/fm] = ", Force);
+      logg[LPropagation].debug("Update momenta: F [GeV/fm] = ", force);
       data.set_4momentum(data.effective_mass(),
-                         data.momentum().threevec() + Force * dt);
+                         data.momentum().threevec() + force * dt);
 
       // calculate the time scale of the change in momentum
-      const double Force_abs = Force.abs();
+      const double Force_abs = force.abs();
       if (Force_abs < really_small) {
         continue;
       }